Limit search frequency.
When the wp-admin side widget fires a search query on a site with a large amount of posts it tends to be slow and cpu intensive.
In this commit we introduce a search minimum search query length of 3. This resolves an issue where short queries (especially single
character) essentially match the entire database. By introducing the query min length with delay the first db query until the user is
more likely to have entered a full word.

We also increase the search handler timeout, to decrease the frequency with which the plugin will hit the db.